Renames WorkerZip -> worker-zip, adds CI validation step
