+++ .eslintignore
@@ -6,5 +6,5 @@
 # The parsers are auto-generated
 /packages/yarnpkg-parsers/sources/grammars/*.js
 
-/packages/yarnpkg-core/sources/ZipWorker/index.js
+/packages/yarnpkg-core/sources/worker-zip/index.js
 /packages/yarnpkg-cli/bundles/yarn.js
