+++ packages/yarnpkg-core/package.json
@@ -58,6 +58,8 @@
     "rollup-plugin-esbuild": "^3.0.2"
   },
   "scripts": {
+    "build:zip:worker": "rollup -c",
+    "update:zip:worker": "run build:zip:worker",
     "postpack": "rm -rf lib",
     "prepack": "run build:compile \"$(pwd)\""
   },
