+++ packages/yarnpkg-core/sources/worker-zip/Worker.ts
