+++ packages/yarnpkg-core/sources/worker-zip/index.d.ts
